Revert "fix(kmp/phase-3): cheapshark storeID is a JSON string, not int"
This reverts commit 2fad74086d5e58569a7609bbb77e4c8cd8b0029a.

+++ remote/cheapshark/src/androidMain/kotlin/pm/bam/gamedeals/remote/cheapshark/mappers/StoreMappers.kt
@@ -14,7 +14,7 @@ internal fun RemoteStore.RemoteStoreImages.toStoreImages(): Store.StoreImages =
 
 internal fun RemoteStore.toStore(): Store =
     Store(
-        storeID = storeID.toInt(),
+        storeID = storeID,
         storeName = storeName,
         isActive = isActive.toBooleanStrict(),
         images = images.toStoreImages()
